x-search/pyproject.toml
[project]
name = "x-search"
version = "0.1.0"
description = "X API search and stream tools for Kalshi market research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "xdk>=0.9.0",
]

[project.scripts]
x-search = "x_search.cli:main"
